Create README.md
The NFT collection, which we published under the name of climate change, still did not receive the interest and demand we expected. The aim of the project is to donate 1 million saplings to one country each year, after obtaining the opportunity to publish on more platforms with the funding made to the collection. In the second stage, we aim to invest in green energy, provide support and funding for projects that will direct our supporters to renewable energy, and increase the accessibility of people to such opportunities and opportunities. Thank you in advance for all your financial support.